feat: finish the evaluation of visdrone dataset

+++ dataset/datasets.py
@@ -121,7 +121,7 @@ class CocoDetection(torchvision.datasets.CocoDetection):
         if "category_id" in batched_target:
             target["labels"] = torch.tensor(batched_target["category_id"])
         else:
-            target["labels"] = torch.empty(0, dtype=torch.int64)
+            target["labels"] = torch.empty(0, dtype=torch.int32)
 
         return image, target
 
@@ -178,7 +178,7 @@ class VisDroneDetection(Dataset):
             ann = {"image_id": image_id}
             boxes = torch.empty([0, 4], dtype=torch.float32)
             scores = torch.empty([0], dtype=torch.float32)
-            labels, truncations, occlusions = [torch.empty([0], dtype=torch.int64) for _ in range(3)]
+            labels, truncations, occlusions = [torch.empty([0], dtype=torch.int32) for _ in range(3)]
 
             file_path = os.path.join(self.ann_root, image_id + ".txt")
             with open(file_path, 'r') as file:
@@ -188,10 +188,10 @@ class VisDroneDetection(Dataset):
                     data = line.strip().split(',')[:8]
                     left, top, width, height, score, category, truncation, occlusion = [float(val) for val in data]
                     boxes = torch.cat([boxes, torch.tensor([[left, top, width, height]], dtype=torch.float32)])
-                    labels = torch.cat([labels, torch.tensor([category], dtype=torch.int64)])
+                    labels = torch.cat([labels, torch.tensor([category], dtype=torch.int32)])
                     scores = torch.cat([scores, torch.tensor([score], dtype=torch.float32)])
-                    truncations = torch.cat([truncations, torch.tensor([truncation], dtype=torch.int64)])
-                    occlusions = torch.cat([occlusions, torch.tensor([occlusion], dtype=torch.int64)])
+                    truncations = torch.cat([truncations, torch.tensor([truncation], dtype=torch.int32)])
+                    occlusions = torch.cat([occlusions, torch.tensor([occlusion], dtype=torch.int32)])
 
             ann.update({'boxes': boxes, 'labels': labels, 'scores': scores, 'truncations': truncations,
                         'occlusions': occlusions})
